refactor: move test structure
